parking boy can assist customer to park when there is a parking lot.

+++ src/test/ParkingLotTest.java
@@ -12,7 +12,7 @@ public class ParkingLotTest {
         ParkingLot parkingLot = new ParkingLot(100);
         Car car = new Car();
 
-        int token = parkingLot.parking(car);
+        Long token = parkingLot.parking(car);
 
         assertThat(car, sameInstance(parkingLot.pick(token)));
     }
@@ -24,7 +24,7 @@ public class ParkingLotTest {
         parkingLot.parking(existedCar);
 
         Car myCar = new Car();
-        int tokenOfMyCar = parkingLot.parking(myCar);
+        Long tokenOfMyCar = parkingLot.parking(myCar);
 
         assertThat(myCar, sameInstance(parkingLot.pick(tokenOfMyCar)));
     }
@@ -33,7 +33,7 @@ public class ParkingLotTest {
     public void should_able_to_pick_up_my_car_after_another_car_parking_given_there_is_a_car_in_parking_lot() {
         ParkingLot parkingLot = new ParkingLot();
         Car myCar = new Car();
-        Integer myToken = parkingLot.parking(myCar);
+        Long myToken = parkingLot.parking(myCar);
 
         Car anotherCar = new Car();
 
@@ -47,7 +47,7 @@ public class ParkingLotTest {
         parkingLot.parking(new Car());
 
         Car car = new Car();
-        Integer nullToken = parkingLot.parking(car);
+        Long nullToken = parkingLot.parking(car);
 
         assertNull(nullToken);
     }
@@ -57,11 +57,11 @@ public class ParkingLotTest {
         ParkingLot parkingLot = new ParkingLot(1);
 
         Car car = new Car();
-        Integer token = parkingLot.parking(car);
+        Long token = parkingLot.parking(car);
         parkingLot.pick(token);
 
         Car myCar = new Car();
-        Integer myToken = parkingLot.parking(myCar);
+        Long myToken = parkingLot.parking(myCar);
 
         assertThat(myCar, sameInstance(parkingLot.pick(myToken)));
 
@@ -71,12 +71,12 @@ public class ParkingLotTest {
     public void should_able_to_park_car_when_another_car_picked_given_parking_lot_is_full() {
         ParkingLot parkingLot = new ParkingLot(1);
 
-        Integer token = null;
+        Long token = null;
         token = parkingLot.parking(new Car());
         parkingLot.pick(token);
 
         Car car = new Car();
-        Integer myToken = parkingLot.parking(car);
+        Long myToken = parkingLot.parking(car);
 
         assertNotNull(myToken);
     }
@@ -85,7 +85,7 @@ public class ParkingLotTest {
     public void should_not_able_to_pick_up_car_when_I_not_park_a_car() {
         ParkingLot parkingLot = new ParkingLot(100);
 
-        Integer token = 2313;
+        Long token = 2313L;
 
         assertNull(parkingLot.pick(token));
     }
@@ -95,12 +95,10 @@ public class ParkingLotTest {
         ParkingLot parkingLot = new ParkingLot(100);
 
         Car car = new Car();
-        Integer myToken = parkingLot.parking(car);
+        Long myToken = parkingLot.parking(car);
 
         assertThat(car, sameInstance(parkingLot.pick(myToken)));
 
         assertNull(parkingLot.pick(myToken));
     }
-
-
 }
